+++ dev/user.clj
@@ -1,5 +1,0 @@
-(ns user
-  (:require
-    [bract.core.dev   :refer [start stop verbose config-files context-file]]
-    [bract.dev.reload :refer [go reinit reset restart]]))
-